requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
django==3.2.25
    # via
    #   -r requirements.in
    #   django-extensions
django-extensions==3.2.3
    # via -r requirements.in
django-nose==1.4.7
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
nose==1.3.7
    # via django-nose
packaging==24.1
    # via gunicorn
psycopg2-binary==2.9.10
    # via -r requirements.in
pytz==2024.2
    # via django
simplejson==3.19.3
    # via -r requirements.in
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via asgiref
